### 2019-08-29 #1687 @whedon
[PRE REVIEW]: kmeRs : K-Mers Similarity Score Matrix
**Submitting author:** @RafalUrniaz (<a href="http://orcid.org/0000-0003-0192-2165">Rafal Urniaz</a>)
**Repository:** <a href="https://github.com/RafalUrniaz/kmeRs" target ="_blank">https://github.com/RafalUrniaz/kmeRs</a>
**Version:** 1.1.0
