### 2019-08-29 #1687 @labarba
👋 @RafalUrniaz — I am the Associate Editor-in-Chief on rotation this week. I'm going to send this back to you for some more work, because the paper does not meet [our expectations](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain). The paper, as submitted, is too brief and doesn't describe the software or its need and uses in research. While we don't expect a full article, it should be 250–1000 words. These items need to be included:

- the summary describing high-level functionality and purpose of the software for a diverse, non-specialist audience;
- a clear statement of need that illustrates the research application of the software (including references to other software meeting similar needs).

TL;DR: the article needs to be improved with more description of how the software works, some examples, and use cases.
