+++ paper/paper.bib
@@ -23,3 +23,83 @@
     year={2018},
     month={Aug}
   }
+
+@article{NEEDLEMAN01,
+    title = "A general method applicable to the search for similarities in the amino acid sequence of two proteins",
+    journal = "Journal of Molecular Biology",
+    volume = "48",
+    number = "3",
+    pages = "443 - 453",
+    year = "1970",
+    issn = "0022-2836",
+    doi = "10.1016/0022-2836(70)90057-4",
+    url = "http://www.sciencedirect.com/science/article/pii/0022283670900574",
+    author = "Saul B. Needleman and Christian D. Wunsch"
+  }
+
+@article{SMITH01,
+    title = "Identification of common molecular subsequences",
+    journal = "Journal of Molecular Biology",
+    volume = "147",
+    number = "1",
+    pages = "195 - 197",
+    year = "1981",
+    issn = "0022-2836",
+    doi = "10.1016/0022-2836(81)90087-5",
+    url = "http://www.sciencedirect.com/science/article/pii/0022283681900875",
+    author = "T.F. Smith and M.S. Waterman"
+  }
+
+@article{DESANTIS01,
+    title = "Simrank: Rapid and sensitive general-purpose k-mer search tool",
+    journal = "BMC ecology",
+    volume = "11",
+    number = "11",
+    pages = "11 - 11",
+    year = "2011",
+    issn = "1472-6785",
+    doi = "10.1186/1472-6785-11-11",
+    url = "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3097142",
+    author = "DeSantis, Todd Z et al."
+  }
+
+@misc{GENOME,
+   author = "{Wikipedia contributors}",
+   title = "Genome",
+   year = "2019",
+   url = "https://en.wikipedia.org/wiki/Genome",
+   note = "[Online; accessed 14-September-2019]"
+ }
+
+@misc{GENE,
+   author = "{Wikipedia contributors}",
+   title = "Gene",
+   year = "2019",
+   url = "https://en.wikipedia.org/wiki/Gene",
+   note = "[Online; accessed 14-September-2019]"
+ }
+
+@misc{Whatisagene,
+   author = "{U.S. National Library of Medicine}",
+   title = "What is a gene?",
+   year = "2019",
+   url = "https://ghr.nlm.nih.gov/primer/basics/gene",
+   note = "[Online; accessed 14-September-2019]"
+ }
+
+@misc{ss,
+   author = "{National Center for Biotechnology Information (NCBI)}",
+   title = "The Statistics of Sequence Similarity Scores",
+   year = "2019",
+   url = "https://www.ncbi.nlm.nih.gov/BLAST/tutorial/Altschul-1.html",
+   note = "[Online; accessed 14-September-2019]"
+ }
+
+@article{JEH01,
+    title = "SimRank: A Measure of Structural-Context Similarity",
+    journal = "Stanford InfoLab Publication Server",
+    pages = "1 - 11",
+    year = "2001",
+    url = "https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3097142",
+    author = "G. Jeh and J. Widom"
+  }
