### 2019-09-14 #1687 @RafalUrniaz
Hi @labarba,

Thank you for your essential comments. I am happy to announce that the paper was revised and updated as you suggested above. Please, find the new one in the paper folder at kmeRs git. I am looking forward to you suggestions after revision.

Best wishes,
Rafal.
