### 2019-09-17 #1687 @kthyng
Now that those changes have been made, it looks like we can move forward.

@csoneson could you edit this submission?
