### 2019-09-17 #1687 @csoneson
@coughls, @VivekTodur, @jlincbio - would two of you be willing to review this submission for JOSS?
