### 2019-09-18 #1687 @VivekTodur
Hello,

Yes, I would be interested.

Regards,
Vivek

On Wed, 18 Sep 2019, 1:11 am Charlotte Soneson, <notifications@github.com>
wrote:

> @coughls <https://github.com/coughls>, @VivekTodur
> <https://github.com/VivekTodur>, @jlincbio <https://github.com/jlincbio>
> - would two of you be willing to review this submission for JOSS?
>
