### 2019-09-18 #1744 @jlincbio
I have some comments about the software and documentation - can I just paste them here?
