### 2019-09-18 #1744 @jlincbio
Comments:

### Functionality
#### Installation
* The `devtools` package in R is not typically installed by default; I would suggest users to install via a tarball release using `R CMD INSTALL` instead, or add a comment about `devtools`.
* I am not sure where this comment really should go, but the two dependencies, `tcR` and `rDNAse`, require way too many other dependencies that are unrelated to the core function of this package. Looking at the R code it seems like only two functions `rDNAse::twoSeqSim()` and `tCR::generate.kmers()` are invoked, and neither of which require calls to features not included outside core R and `Biostrings`, perhaps incorporate these functions into the package and cite the two packages instead?
Sources: [`rDNAse::twoSeqSim`](https://rdrr.io/github/wind22zhu/rDNAse/) and [`tcR::generate.kmers`](https://github.com/imminfo/tcr/blob/master/R/kmers.R)
* I got an error on a fresh Ubuntu 18.04 LTS/R Open 3.5.3 installation while installing the `tcR` dependency `igraph` about missing gfortran; this is more or less carelessness on my end but also a side effect of requiring unnecessary dependencies.

### Documentation
#### Statement of need
While the documentation describes the purpose of the tool as to "calculate similarity score matrix for DNA k-mers," there is no description on what problems this tool is designed to solve or the target audience.
#### Functionality documentation
There is no documentation or tutorial on how to export results from the R package to use with the companion Python program (heatmap4kmers), even though Figure 1 is actually generated from the Python program and the R package includes an export function; from my perspective it seems like a core feature that is left undocumented.
#### Example usage
The documentation includes a clear walkthrough, but it would be nice to have a more real-world scenario use case than GATTACA.
#### Community guidelines
This is missing in the documentation but can be fixed relatively quickly.

### Software paper
#### Summary
since the introduction discusses a layman’s description of DNA, so technically it satisfies the requirement that it describes the tool to a _"diverse, non-specialist audience"_; however, it needs a little more discussion on the purpose of this tool in the introduction section in my opinion.
#### Statement of need
The paper describes the purpose behind creating such a tool to generate similarity score matrices but does not describe the sort of research problem this tool can help address; while similarity scores are calculated to compare sequences, from time to time this can be done without using this tool, so I suggest that the paper should try to address the purpose of this tool better. Since this is technically an academic paper, it should adhere to JOSS submission requirements that require that _"the software should have an obvious research application"_ and _"be a significant contribution to the available open source software."_
#### State of the field
This is entirely missing in the paper. A discussion on why someone would bother to calculate these scores to compare sequence similarities would be a starting point to illustrate the purpose of this tool.
#### Quality of writing
It is a bit unclear as it stands what the purpose and use cases of this tool are. The manuscript describes what some the basic keywords mean, but fails to address the scientific idea. For instance, the author mentions a list of PAM and BLOSUM matrices frequently used in bioinformatics, and while the different use cases for these matrices are too fine and beyond the scope of this manuscript, merely listing them and not discussing what they are clutters the writing.
#### References
The paper cities Wikipedia and PLoSWiki entries; this I find a bit problematic for archival purposes. Additionally, the NCBI citations are not archived either.

### Miscellaneous
* Personally I would like to see the functionality of plotting the similarity heatmap (Figure 1) be incorporated into the R package, either via a command call to invoke the Python software or native R code (ideally, either via the native `heatmap()` function or `ggplots` (one of the dependencies installed by `tcR`), seeing how the paper includes this figure and in a way counts the ability to generate data for this tool as part of the feature set. Without this, the package seems a bit bare bones, since it can be construed as a wrapper for `Biostrings::pairwiseAlignment`; this technically goes against the requirement that minor utility packages being unacceptable for submission.
* The tutorial uses BLOSUM62 with DNA sequences; in this situation I suggest using amino acid sequences instead.
