### 2019-10-03 #1744 @jlincbio
@csoneson thanks for checking in. Please let me know when I need to follow up. I have a few things on the agenda for the next two weeks but quick checks are not a problem.
