### 2019-10-03 #1744 @VivekTodur
Yes, I am almost done with the review. I can revert back by this Sunday...

Thanks

On Thu, 3 Oct 2019, 12:27 am Charlotte Soneson, <notifications@github.com>
wrote:

> Hi all, just checking in on how the reviews are progressing here -
> @jlincbio <https://github.com/jlincbio>, thanks for your comments above!
> @VivekTodur <https://github.com/VivekTodur> - do you think you will have
> the chance to provide your review in the coming week? Thanks!
>
