### 2019-10-11 #1744 @VivekTodur
@csoneson Thanks for the opportunity to review this manuscript and my apologies for the delay. Below are my comments,

**Functionality**
**Installation**
Installation is quick and simple as described in the documentation on Ubuntu 19.04. Shouldn't be a problem in other versions of Linux computers that are prepared for biological sequence analysis.

**Documentation**
**Statement of need**
Manual for the tool "kmeRs: K-Mers Similarity Score Matrix" is not clearly describing the general purpose and requirement of the tool, A quick note is required.

**Functionality documentation**
Various functions of the tool are neatly illustrated with relevant examples, but it lacks documentation on how to export the results and it's output format. Also, manuscript containings Kmer Heatmap image generated from heatmap4kmers, A quick note on this is very useful.

**Example usage**
The documentation includes detailed instructions on the How-To-Use. As mentioned earlier output export method is missing

**Community guidelines**
General guidelines are missing, should be added.

**Software paper**
**Summary**
The tool is developed to address problems related to biological sequence similarity searches. In this, very focused context author has neatly explained the purpose.

**Statement of need**
Paper titled "K-Mers Similarity Score Matrix" describes the purpose and requirement of the tool in layman terms. The author needs to describe the research object, a real-world requirement and how kmeRs is outperforming the other tools.

**State of the field**
An author should include this section as it is missing. Also, it would be nice to see a quick note on how this tool can be integrated into existing other pipelines.

**Quality of writing**
There are plenty of tools/algorithms already available to solve the sequence similarity search problem. The author needs to explain the exact purpose of kmeRs, and how it is different from the existing one in terms of functionalities. A computational benchmark [CPU vs Memory] would add more value to the manuscript. Empirical formula to calculate KmeR Score is also missing in the manuscript.

**References**
Considering this is software paper, including citations from the scientific journal, are sufficient. Those Wikipedia citations can be omitted.
