### 2019-10-19 #1744 @csoneson
@VivekTodur - thanks for your review!
@RafalUrniaz - please go through the comments from the reviewers above, and report back here when you are ready for them to take another look. Of course, don't hesitate to ping me or the reviewers if you have questions.
