### 2019-10-30 #1744 @csoneson
Hi @RafalUrniaz - I just wanted to make sure that you have seen the comments above from the reviewers, and check whether you would have an estimate of when you might be able to have a new version for them to consider. Please let us know if you have questions.
