### 2020-01-23 #1744 @csoneson
I have been in contact with @RafalUrniaz via email, and he would like to withdraw this submission since he will not have time to get back to it in the near future.

@openjournals/joss-eics - I will close this issue and add the 'withdrawn' label, is there anything else that should be done from my side?

@jlincbio, @VivekTodur - thanks a lot for your reviews, really appreciate your time and input.
